requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
astroid==2.15.8
attrs==23.2.0
Babel==2.15.0
certifi==2024.2.2
et-xmlfile==1.1.0
frozenlist==1.4.1
i18n==0.2
idna==3.7
lazy-object-proxy==1.10.0
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
py==1.11.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
redis-dict==2.5.0
semver==3.0.2
six==1.16.0
toml==0.10.2
typing_extensions==4.12.0
tzdata==2024.1
wrapt==1.16.0
yarl==1.9.4
